+++ core/src/main/java/org/apache/calcite/rel/rules/JoinExtractFilterRule.java
@@ -16,11 +16,7 @@
  */
 package org.apache.calcite.rel.rules;
 
-import org.apache.calcite.plan.RelOptRule;
-import org.apache.calcite.plan.RelOptRuleCall;
-import org.apache.calcite.rel.RelNode;
 import org.apache.calcite.rel.core.Join;
-import org.apache.calcite.rel.core.JoinRelType;
 import org.apache.calcite.rel.core.RelFactories;
 import org.apache.calcite.rel.logical.LogicalJoin;
 import org.apache.calcite.tools.RelBuilderFactory;
@@ -39,7 +35,7 @@ import org.apache.calcite.tools.RelBuilderFactory;
  * {@link org.apache.calcite.rel.core.Join}, not just
  * {@link org.apache.calcite.rel.logical.LogicalJoin}.</p>
  */
-public final class JoinExtractFilterRule extends RelOptRule {
+public final class JoinExtractFilterRule extends AbstractJoinExtractFilterRule {
   //~ Static fields/initializers ---------------------------------------------
 
   /** The singleton. */
@@ -50,58 +46,15 @@ public final class JoinExtractFilterRule extends RelOptRule {
   //~ Constructors -----------------------------------------------------------
 
   /**
-   * Creates an JoinExtractFilterRule.
-   */
-  @Deprecated // to be removed before 2.0
-  public JoinExtractFilterRule(Class<? extends Join> clazz) {
-    this(clazz, RelFactories.LOGICAL_BUILDER);
-  }
-
-  /**
    * Creates a JoinExtractFilterRule.
    */
   public JoinExtractFilterRule(Class<? extends Join> clazz,
       RelBuilderFactory relBuilderFactory) {
-    super(operand(clazz, any()), relBuilderFactory, null);
+    super(operand(clazz, any()), relBuilderFactory, RelFactories.DEFAULT_FILTER_FACTORY);
   }
 
   //~ Methods ----------------------------------------------------------------
 
-  public void onMatch(RelOptRuleCall call) {
-    final Join join = call.rel(0);
-
-    if (join.getJoinType() != JoinRelType.INNER) {
-      return;
-    }
-
-    if (join.getCondition().isAlwaysTrue()) {
-      return;
-    }
-
-    if (!join.getSystemFieldList().isEmpty()) {
-      // FIXME Enable this rule for joins with system fields
-      return;
-    }
-
-    // NOTE jvs 14-Mar-2006:  See JoinCommuteRule for why we
-    // preserve attribute semiJoinDone here.
-
-    RelNode cartesianJoinRel =
-        join.copy(
-            join.getTraitSet(),
-            join.getCluster().getRexBuilder().makeLiteral(true),
-            join.getLeft(),
-            join.getRight(),
-            join.getJoinType(),
-            join.isSemiJoinDone());
-
-    final RelFactories.FilterFactory factory =
-        RelFactories.DEFAULT_FILTER_FACTORY;
-    RelNode filterRel =
-        factory.createFilter(cartesianJoinRel, join.getCondition());
-
-    call.transformTo(filterRel);
-  }
 }
 
 // End JoinExtractFilterRule.java
